+++ eNMS/database.py
@@ -544,6 +544,28 @@ class Database:
             for instance in self.fetch_all(model, **kwargs)
         ]
 
+    def try_commit(self, transaction, *args, **kwargs):
+        for index in range(self.retry_commit_number):
+            try:
+                result = transaction(*args, **kwargs)
+                self.session.commit()
+                break
+            except Exception as exc:
+                self.session.rollback()
+                if index == self.retry_commit_number - 1:
+                    error(f"Commit n°{index + 1} failed ({format_exc()})")
+                    raise exc
+                else:
+                    warning(f"Commit n°{index + 1} failed ({str(exc)})")
+                sleep(self.retry_commit_time * (index + 1))
+        return result
+
+    def try_set(self, instance, property, value):
+        def transaction():
+            setattr(instance, property, value)
+
+        self.try_commit(transaction)
+
     def factory(
         self, _class, commit=False, no_fetch=False, rbac="edit", user=None, **kwargs
     ):
@@ -576,19 +598,7 @@ class Database:
         if not commit:
             instance = transaction(_class, **kwargs)
         else:
-            for index in range(self.retry_commit_number):
-                try:
-                    instance = transaction(_class, **kwargs)
-                    self.session.commit()
-                    break
-                except Exception as exc:
-                    self.session.rollback()
-                    if index == self.retry_commit_number - 1:
-                        error(f"Commit n°{index} failed ({format_exc()})")
-                        raise exc
-                    else:
-                        warning(f"Commit n°{index} failed ({str(exc)})")
-                    sleep(self.retry_commit_time * (index + 1))
+            instance = self.try_commit(transaction, _class, **kwargs)
         return instance
 
     def get_credential(