+++ eNMS/models/administration.py
@@ -59,6 +59,7 @@ class Worker(AbstractBase):
     __tablename__ = type = class_type = "worker"
     id = db.Column(Integer, primary_key=True)
     name = db.Column(db.SmallString, unique=True)
+    process_id = db.Column(Integer)
     description = db.Column(db.LargeString)
     subtype = db.Column(db.TinyString)
     last_update = db.Column(db.TinyString)
@@ -298,8 +299,10 @@ class File(AbstractBase):
 
     def update(self, move_file=True, **kwargs):
         old_path = self.full_path
+        self.full_path = f"{vs.file_path}{kwargs['path']}"
+        if not str(Path(self.full_path).resolve()).startswith(f"{vs.file_path}/"):
+            raise Exception("The path resolves outside of the files folder.")
         super().update(**kwargs)
-        self.full_path = f"{vs.file_path}{self.path}"
         if exists(str(old_path)) and not exists(self.full_path) and move_file:
             move(old_path, self.full_path)
         self.name = self.path.replace("/", ">")
@@ -356,9 +359,9 @@ class Folder(File):
 
     def __init__(self, **kwargs):
         full_path = f"{vs.file_path}{kwargs['path']}"
+        self.update(**kwargs)
         if not exists(full_path) and not kwargs.get("migration_import"):
             makedirs(full_path)
-        self.update(**kwargs)
 
 
 class Secret(AbstractBase):