+++ eNMS/models/automation.py
@@ -463,10 +463,12 @@ class Run(AbstractBase):
     def run(self):
         worker = db.factory(
             "worker",
-            name=str(getpid()),
+            name=f"{vs.server} - {getpid()}",
+            process_id=getpid(),
             subtype=environ.get("_", "").split("/")[-1],
             server_id=vs.server_id,
             rbac=None,
+            commit=True,
         )
         self.worker = worker
         vs.run_targets[self.runtime] = set(