+++ eNMS/static/js/table.js
@@ -16,6 +16,7 @@ import {
   loadTypes,
   notify,
   openPanel,
+  sanitize,
   serializeForm,
   showChangelogPanel,
   showConfirmationPanel,
@@ -44,6 +45,7 @@ export class Table {
     this.columns.forEach((column) => {
       if (visibleColumns) column.visible = visibleColumns.includes(column.data);
       column.name = column.data;
+      if (!column.html) column.render = $.fn.dataTable.render.text();
     });
     this.userFiltering = localStorage.getItem(`userFiltering-${this.type}`) || "users";
     this.id = `${this.type}${id ? `-${id}` : ""}`;
@@ -677,7 +679,8 @@ tables.device = class DeviceTable extends Table {
 tables.network = class NetworkTable extends Table {
   addRow(kwargs) {
     let row = super.addRow(kwargs);
-    row.name = `<b><a href="/network_builder/${row.path}">${row.name}</a></b>`;
+    const rowName = sanitize(row.name);
+    row.name = `<b><a href="/network_builder/${row.path}">${rowName}</a></b>`;
     row.links = `<b><a href="#" onclick="eNMS.table.displayRelationTable(
       'link', ${row.instance}, {parent: '${this.id}', from: 'networks',
       to: 'links'})">Links</a></b>`;
@@ -1019,11 +1022,11 @@ tables.service = class ServiceTable extends Table {
   addRow(kwargs) {
     let row = super.addRow(kwargs);
     if (row.type == "workflow") {
-      row.name = `<b><a href="/workflow_builder/${row.path}">${row.name}</a></b>`;
+      row.name = `<b><a href="/workflow_builder/${row.path}">${sanitize(row.name)}</a></b>`;
     } else if (!row.shared) {
       row.name = row.name.replace(
         row.scoped_name,
-        `<b><a href="/workflow_builder/${row.workflow_path}">${row.scoped_name}</a></b>`
+        `<b><a href="/workflow_builder/${row.workflow_path}">${sanitize(row.scoped_name)}</a></b>`
       );
     }
     for (const model of ["device", "pool"]) {
@@ -1497,7 +1500,6 @@ tables.group = class GroupTable extends Table {
       this.columnDisplay(),
       this.displayChangelogButton(),
       this.refreshTableButton(),
-      this.bulkFilteringButton(),
       this.clearSearchButton(),
       this.copyTableButton(),
       this.createNewButton(),